newfile git way

+++ 04_04diff_of_type_in_print_scanf[0404]/0404excer.c
@@ -2,16 +2,16 @@
 
 int main () {
 
-    float num1, num2, num3;
+    int num1, num2, num3;
 
     printf("請輸入第一個整數\t");
-    scanf("%f" , &num1);
+    scanf("%d" , &num1);
     printf("請輸入第二個整數\t");
-    scanf("%f" , &num2);
+    scanf("%d" , &num2);
     printf("請輸入第三個整數\t");
-    scanf("%f" , &num3);
+    scanf("%d" , &num3);
 
-    int avg = (num1 + num2 + num3) / 3.0;
+    double avg = (num1 + num2 + num3) / 3.;
     printf("平均： %f\n ", avg);
     return 0 ;
 }
